Using a pythonic project structure
