+++ TUIFIManager/settings.py
@@ -1,0 +1,19 @@
+import unicurses as uc
+
+PADDING_LEFT   = 2
+PADDING_RIGHT  = 2
+PADDING_TOP    = 1
+PADDING_BOTTOM = 0
+
+
+MY_COLOR_PAIRS = (
+    (uc.COLOR_WHITE  ,uc.COLOR_BLACK  ),
+    (uc.COLOR_YELLOW ,uc.COLOR_BLACK  ),
+    (uc.COLOR_RED    ,uc.COLOR_BLACK  ),
+    (uc.COLOR_BLUE   ,uc.COLOR_BLACK  ),
+    (uc.COLOR_GREEN  ,uc.COLOR_BLACK  ),
+    (uc.COLOR_BLACK  ,uc.COLOR_WHITE  ),
+    (uc.COLOR_BLUE   ,uc.COLOR_WHITE  ),
+    (uc.COLOR_CYAN   ,uc.COLOR_BLACK  ),
+    (uc.COLOR_BLACK  ,uc.COLOR_YELLOW ),
+)
